restructuring test files
